linebot-20250101/requirements.txt
line-bot-sdk==2.3.0
google-cloud-firestore==2.19.0
firebase-admin==6.6.0
